requirements.txt
azure-functions
azure-search-documents
azure-core
openai
python-dotenv
